+++ ceno_zkvm/src/utils.rs
@@ -225,7 +225,7 @@ pub fn add_mle_list_by_expr<'a, E: ExtensionField>(
     challenges: &[E],
     // sumcheck batch challenge
     alpha: E,
-) -> BTreeSet<u16> {
+) -> BTreeSet<u32> {
     assert!(expr.is_monomial_form());
     let monomial_terms = expr.evaluate(
         &|_| unreachable!(),
@@ -287,7 +287,7 @@ pub fn add_mle_list_by_expr<'a, E: ExtensionField>(
     monomial_terms
         .into_iter()
         .flat_map(|(_, monomial_term)| monomial_term.into_iter().collect_vec())
-        .collect::<BTreeSet<u16>>()
+        .collect::<BTreeSet<u32>>()
 }
 
 #[cfg(test)]
